requirements.txt
streamlit
streamlit-extras
google-generativeai
python-dotenv
PyPDF2
pillow